# server/api.yaml
openapi: 3.0.3
info:
  title: Hospital Management System API
  version: 1.0.0
  description: |
    Role-based API for Hospital Management System (HMS).
    Authentication uses JWT Bearer tokens.
servers:
  - url: http://127.0.0.1:5000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/doctor/me:
    get:
      tags: [Doctor]
      summary: Doctor profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Doctor details
  /api/doctor/availability:
    get:
      tags: [Doctor]
      summary: List doctor availability
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Availability list
    post:
      tags: [Doctor]
      summary: Create availability slot
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Availability created
  /api/doctor/availability/{slot_id}:
    delete:
      tags: [Doctor]
      summary: Delete availability slot
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: slot_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Slot deleted
  /api/doctor/appointments:
    get:
      tags: [Doctor]
      summary: List doctor appointments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Appointment list
  /api/doctor/appointments/{appointment_id}:
    patch:
      tags: [Doctor]
      summary: Update appointment status/treatment details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: appointment_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Appointment updated
  /api/doctor/patient-history/{patient_id}:
    get:
      tags: [Doctor]
      summary: Completed history for assigned patient
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Patient history
  /api/patient/me:
    get:
      tags: [Patient]
      summary: Patient profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Patient profile
    patch:
      tags: [Patient]
      summary: Update patient profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Patient updated
...
  /api/patient/doctors/{doctor_id}:
    get:
      tags: [Patient]
      summary: Get doctor details
      parameters:
        - in: path
          name: doctor_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Doctor details
  /api/patient/doctors/{doctor_id}/availability:
    get:
      tags: [Patient]
      summary: Get doctor availability by date
      parameters:
        - in: path
          name: doctor_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Availability list
  /api/patient/appointments:
    get:
      tags: [Patient]
      summary: List patient appointments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Appointment list
    post:
      tags: [Patient]
      summary: Book appointment
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Appointment booked
  /api/patient/appointments/{appointment_id}:
    patch:
      tags: [Patient]
      summary: Cancel appointment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: appointment_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Appointment updated
  /api/patient/exports/treatments:
    post:
      tags: [Patient]
      summary: Trigger async treatment CSV export
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Export task accepted
  /api/patient/exports/{task_id}:
    get:
      tags: [Patient]
      summary: Poll export status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Export status